Orientation & Terms: * Added terms of use component. * Added orientation component with a sample video. * Updated global CSS. * Added survey component (TO DO)
