+++ apps/conte-pwa/src/app/app-routing.module.ts
@@ -2,8 +2,9 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { SplashScreenComponent } from './components/splash-screen/splash-screen.component';
 import { AuthenticationComponent } from './components/authentication/authentication.component';
-import { TermsOfUseComponent } from './components/termsOfUse/terms-of-use/terms-of-use.component';
-import { OrientationVideoComponent } from './components/shared/orientation/orientation-video/orientation-video.component';
+import { TermsOfUseComponent } from './components/terms-of-use/terms-of-use.component';
+import { OrientationComponent } from './components/orientation/orientation.component';
+import { SurveyComponent } from './components/survey/survey.component';
 
 const routes: Routes = [
   {
@@ -14,7 +15,8 @@ const routes: Routes = [
   { path: '', component: SplashScreenComponent },
   { path: 'authentication', component: AuthenticationComponent },
   { path: 'terms', component: TermsOfUseComponent },
-  { path: 'orientation', component: OrientationVideoComponent },
+  { path: 'orientation', component: OrientationComponent },
+  { path: 'survey', component: SurveyComponent },
 ];
 
 @NgModule({
