+++ apps/conte-pwa/src/app/components/orientation/orientation.component.html
@@ -1,0 +1,33 @@
+<div class="container-fluid d-flex orientation-screen">
+  <div class="container-fluid d-flex orientation-container">
+    <h1 class="display-3">Conte PWA</h1>
+    <h1 class="display-6 mb-5">ORIENTATION</h1>
+    <!-- Orientation Video -->
+    <video width="100%" height="100%" class="mt-5 mb-5" autoplay muted controls>
+      <source src="../../../assets/orientation/video/orientation.mp4" type="video/mp4" />
+      Your browser does not support the video tag.
+    </video>
+
+    <!-- Checkbox -->
+    <div class="d-flex terms-check mt-3 mb-3">
+      <input type="checkbox" class="me-2" [(ngModel)]="acceptTermsCheck" />
+      <span class="text" (click)="acceptTermsCheck = !acceptTermsCheck">I understand how the app works.</span>
+    </div>
+
+    <!-- Confirmation -->
+    <div class="d-flex justify-content-center mt-4 mb-3">
+      <button
+        class="btn btn-primary btn-confirm"
+        type="button"
+        (click)="acceptTerms()"
+        [disabled]="!acceptTermsCheck || buttonState == 'loading'"
+      >
+        <span *ngIf="buttonState == 'static'">Proceed</span>
+        <div *ngIf="buttonState == 'loading'">
+          <span class="spinner-border spinner-border-sm me-2" role="status" aria-hidden="true"></span>
+          <span>Confirming your orientation...</span>
+        </div>
+      </button>
+    </div>
+  </div>
+</div>
