+++ apps/conte-pwa/src/app/components/orientation/orientation.component.scss
@@ -1,0 +1,53 @@
+.orientation-screen {
+  height: 100vh;
+  background: radial-gradient(ellipse at right bottom, rgb(50, 51, 62) 0%, rgb(5, 107, 25) 100%);
+  padding: 5%;
+  align-items: center;
+}
+
+.orientation-container {
+  justify-content: center;
+  padding: 6% 4% 6% 4%;
+  width: 100%;
+  background-color: white;
+  text-align: center;
+  align-items: center;
+  border-radius: 8px;
+  flex-wrap: wrap;
+
+  h1 {
+    width: 100%;
+  }
+
+  video {
+    border-radius: 6px;
+  }
+
+  .terms-check {
+    justify-content: center;
+    width: 100%;
+  }
+
+  .btn-confirm {
+    padding: 10px;
+    width: 210px;
+    border-radius: 20px;
+    background: radial-gradient(ellipse at right bottom, rgb(50, 51, 62) 0%, rgb(5, 107, 25) 100%);
+    border-style: none;
+    color: white;
+    font-weight: 600;
+  }
+
+  .btn-confirm:hover {
+    background-color: rgba(10, 136, 43, 0.5);
+  }
+}
+
+@media screen and (min-width: 1350px) {
+  .orientation-screen {
+    height: 100%;
+    background: radial-gradient(ellipse at right bottom, rgb(50, 51, 62) 0%, rgb(5, 107, 25) 100%);
+    padding: 5%;
+    align-items: center;
+  }
+}
