--- apps/conte-pwa/src/app/components/shared/orientation/orientation-video/orientation-video.component.html
@@ -1,1 +1,0 @@
-<p>orientation-video works!</p>
