--- apps/conte-pwa/src/app/components/shared/orientation/orientation-video/orientation-video.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'conte-orientation-video',
-  templateUrl: './orientation-video.component.html',
-  styleUrls: ['./orientation-video.component.scss'],
-})
-export class OrientationVideoComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
