+++ apps/conte-pwa/src/app/components/splash-screen/splash-screen.component.html
@@ -1,10 +1,9 @@
 <div class="container-fluid">
   <div class="container-fluid d-flex splash-screen">
     <div class="logo-container d-flex justify-content-center">
-      <img
-        class="logo"
-        src="../../assets/logos/Throwing Progression Pro-02.png"
-      />
+      <h1 class="display-3">Conte PWA</h1>
+      <!-- For Logo -->
+      <!-- <img class="logo" src="../../assets/logos/Throwing Progression Pro-02.png" /> -->
     </div>
     <div class="spinner-border" role="status"></div>
   </div>
