+++ apps/conte-pwa/src/app/components/survey/survey.component.html
@@ -1,0 +1,1 @@
+<p>survey works!</p>
