+++ apps/conte-pwa/src/app/components/survey/survey.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'conte-survey',
+  templateUrl: './survey.component.html',
+  styleUrls: ['./survey.component.scss'],
+})
+export class SurveyComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}
