+++ apps/conte-pwa/src/app/components/terms-of-use/terms-of-use.component.html
@@ -1,0 +1,53 @@
+<div class="container-fluid terms-screen">
+  <div class="container-fluid d-flex terms-container">
+    <h1 class="display-3">Conte PWA</h1>
+    <h1 class="display-5 mb-5">Terms & Conditions</h1>
+    <!-- Terms -->
+    <p>
+      What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has
+      been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and
+      scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into
+      electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of
+      Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus
+      PageMaker including versions of Lorem Ipsum.What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the printing
+      and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an
+      unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five
+      centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in
+      the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop
+      publishing software like Aldus PageMaker including versions of Lorem Ipsum. What is Lorem Ipsum? Lorem Ipsum is
+      simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy
+      text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen
+      book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining
+      essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum
+      passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem
+      Ipsum.What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum
+      has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type
+      and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into
+      electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of
+      Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus
+      PageMaker including versions of Lorem Ipsum.
+    </p>
+
+    <!-- Checkbox -->
+    <div class="d-flex terms-check mt-3 mb-3">
+      <input type="checkbox" class="me-2" [(ngModel)]="acceptTermsCheck" />
+      <span class="text" (click)="acceptTermsCheck = !acceptTermsCheck">I accept the terms & conditions</span>
+    </div>
+
+    <!-- Confirmation -->
+    <div class="d-flex justify-content-center mt-4 mb-3">
+      <button
+        class="btn btn-primary btn-confirm"
+        type="button"
+        (click)="acceptTerms()"
+        [disabled]="!acceptTermsCheck || buttonState == 'loading'"
+      >
+        <span *ngIf="buttonState == 'static'">Proceed</span>
+        <div *ngIf="buttonState == 'loading'">
+          <span class="spinner-border spinner-border-sm me-2" role="status" aria-hidden="true"></span>
+          <span>Setting everything up...</span>
+        </div>
+      </button>
+    </div>
+  </div>
+</div>
