+++ apps/conte-pwa/src/app/components/terms-of-use/terms-of-use.component.ts
@@ -1,0 +1,19 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'conte-terms-of-use',
+  templateUrl: './terms-of-use.component.html',
+  styleUrls: ['./terms-of-use.component.scss'],
+})
+export class TermsOfUseComponent implements OnInit {
+  acceptTermsCheck: boolean = false;
+  buttonState: string = 'static';
+
+  constructor() {}
+
+  ngOnInit(): void {}
+
+  acceptTerms(){
+    this.buttonState = 'loading';
+  }
+}
