--- apps/conte-pwa/src/app/components/termsOfUse/terms-of-use/terms-of-use.component.html
@@ -1,1 +1,0 @@
-<p>terms-of-use works!</p>
