--- apps/conte-pwa/src/app/components/termsOfUse/terms-of-use/terms-of-use.component.spec.ts
@@ -1,22 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { TermsOfUseComponent } from './terms-of-use.component';
-
-describe('TermsOfUseComponent', () => {
-  let component: TermsOfUseComponent;
-  let fixture: ComponentFixture<TermsOfUseComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [TermsOfUseComponent],
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(TermsOfUseComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
