--- apps/conte-pwa/src/app/components/termsOfUse/terms-of-use/terms-of-use.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'conte-terms-of-use',
-  templateUrl: './terms-of-use.component.html',
-  styleUrls: ['./terms-of-use.component.scss'],
-})
-export class TermsOfUseComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
