+++ apps/conte-pwa/src/styles.scss
@@ -13,5 +13,4 @@
 body,
 html {
   width: 100%;
-  height: 100vh;
 }
